service_programacion/requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
pybreaker==1.4.1
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
